changed computation of acc

+++ rotator/createRotation.cpp
@@ -28,8 +28,8 @@ const int lenghtOfEndingOfFile=7;
 using namespace std;
 using namespace pcl;
 
-float getRadFromDegree(int degree){
-  return (float(degree)*M_PI)/180;
+float getRadFromDegree(double degree){
+  return float((float(degree)*M_PI)/180);
 
 }
 
@@ -95,7 +95,7 @@ int main(int argc, char *argv[])
 {
 
 //  int angles[]={-30,-27,-24,-21,-18,-15,-12,-9,-6,-3,0,3,6,9,12,15,18,21,24,27,30};
-  int angles[]={-4.5,-3,-1.5,0,1.5,3,4.5};
+  double angles[]={-4.5,-3,-1.5,0,1.5,3,4.5};
 
   int numOfAngles=sizeof(angles)/sizeof(int);
   int currentAnglesIndex=0;
